Improved naming and split to multiple files
